requirements.txt
PyAudioWPatch>=0.2.12
numpy>=1.24.0
windows-curses>=2.3.1
winrt-Windows.Media.Control>=3.2.0
winrt-Windows.Foundation>=3.2.0
